FEAT - Implement HomeScreen.js Join Game button which triggers joinGame action creator
